Use -std=c++0x
* Use boost built with -std=c++0x
* Use explicit namespaces when using classes that overlap between
      std:: & boost:: when using c++0x :
    - shared_array
    - shared_ptr
    - make_tuple
    - tuple
    - ref
    - cref
    - weak_ptr
    - current_exception

Reviewed-on: http://gerrit.lab.local:8080/4422
Tested-by: build
Smoketests: build

+++ LogManager.cpp
@@ -315,7 +315,9 @@ std::pair<bool, unsigned int> Logfile::GetFileMask(const Forte::FString& file) c
 
     if(i != mFileMasks.end())
     {
-        return make_pair<bool, unsigned int>(true, i->second);
+        return make_pair<bool, unsigned int>(
+            true,
+            static_cast<unsigned int>(i->second));
     }
 
     return make_pair<bool, unsigned int>(false, 0);
